Moved code to controller file
