+++ routes/courses.js
@@ -1,5 +1,6 @@
 var express = require('express');
 var router = express.Router();
+var instructorController = require('../routes/instructorController');
 
 /* List ALL Courses */
 router.route('/courses')
@@ -7,24 +8,7 @@ router.route('/courses')
 
 /* Create Course */
 router.route('/courses')
-  .post(function(req, res) {
-      req.app.locals.db.run("INSERT INTO course_map (course_name, institution, department) VALUES (?,?,?)", [req.body.name, req.body.institution, req.body.department], function(err, row) {
-          if (err) {
-              return res.send({error: "new course could not be created"});
-          }
-          if (!row) {
-              return res.send({error: "new course could not be created"});
-          } else {
-              var response = {}
-              response["id"] = this.lastID;
-              response["name"] = req.body.name;
-              response["institution"] = req.body.institution;
-              response["department"] = req.body.department;
-              return res.send(response);
-          }
-      })
-
-  });
+  .post(instructorController.createCourse);
 
 /* Retrieve a Course */
 router.route('/courses/:courseID')
